requirements.txt
openai
nltk
PyPDF2
python-docx
faiss-cpu
numpy
python-dotenv
tkinter
